+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes will be documented in this file.
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
 ## [Unreleased]
+- Add `SentEnvelopeReport.timestampTsaPostSignumCount`
+- Add `SentEnvelopeReport.timestampTsaIcaCount`
+- Add `SentEnvelopeReport.timestampAtsaIcaCount`
 - Edit `EnvelopeRecipientApproval` resource
 - Add `BulkSignature.completedAt` in to resource