+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Add `EnvelopeTemplateRecipient.birthdate`
 - Add `EnvelopeTemplateRecipient.birthnumber`
 - Add `EnvelopeTemplateDocument.assignments` in to resource
+- Add `MyPreferences.twoFactorAuthStatus` in to resource
 
 ## [2.7.1] - 2025-03-31
 - Update CHANGELOG.md and version in DigiSign.php