+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Add `MyPreferences.twoFactorAuthStatus` in to resource
 - Add `TwoFactorAuthEndpoint` and `AccountMeTwoFactorAuthEndpoint` endpoint
 - Add `AccountMeTwoFactorAuthEndpoint.get` endpoint
+- Add `TwoFactorAuthEndpoint.reset` endpoint
 
 ## [2.7.1] - 2025-03-31
 - Update CHANGELOG.md and version in DigiSign.php